ai-fake-detection/requirements.txt
flask
transformers
torch
pillow
opencv-python-headless
